feat: relocate remove_sudo migration code
