+++ runtime/litmus/src/lib.rs
@@ -130,7 +130,6 @@ pub type Executive = frame_executive::Executive<
 	// it was reverse order before.
 	// See the comment before collation related pallets too.
 	AllPalletsWithSystem,
-	migration::RemoveSudoAndStorage<Runtime>,
 >;
 
 impl_opaque_keys! {
