+++ runtime/litmus/src/migration/migration.md
@@ -1,0 +1,6 @@
+P9115.rs
+https://github.com/litentry/litentry-parachain/releases/tag/v0.9.11-1
+ # This migration is for the remove of Sudo on Litmus
+
+ <!-- RemoveSudoAndStorage -->
+ # The main purpose of runtime upgrade is for removing sudo and its storage.
